kasih tombol hapus pada semua modul, perbaikan pada pengadaan, dan pemberian fitur login

+++ app/Http/Controllers/PengadaanController.php
@@ -6,9 +6,17 @@ use App\Pengadaan;
 use App\MitraKerja;
 use Illuminate\Http\Request;
 use App\Gudang;
+use Validator;
+use DB;
 
 class PengadaanController extends Controller
 {
+
+    // public function __construct()
+    // {
+    //     DB::enableQueryLog();
+    // }
+
     /**
      * Display a listing of the resource.
      *
@@ -51,22 +59,38 @@ class PengadaanController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'nama'=>'required|min:3', 
-            'bidang'=>'required|min:3', 
-            'kontak'=>'required|min:3', 
+        $rules = [];
+        $validator = Validator::make($request->all(), [
+            'tanggal'=>'required|date_format:Y-m-d', 
+            'jumlah'=>'required|numeric', 
+            'biaya'=>'required|numeric', 
+            'biaya_transport'=>'required|numeric',
         ]);
+        if(!isset($request->id_gudang)){
+            return redirect()->back()->withErrors($validator)->withInput()->with('error_msg', 'Setidaknya pilih salah satu gudang');
+        }
+        foreach ($request->id_gudang as $id_gudang) {
+            $rules['isi_gudang_'.$id_gudang] = 'required|numeric';
+        }
+        $request->validate($rules);
         if(Pengadaan::count() == 0){
             DB::statement('set foreign_key_checks=0;');
             Pengadaan::truncate();
         }
         $pengadaan = Pengadaan::create([
-            'nama'=>$request->nama,
-            'bidang'=>$request->bidang,
-            'kontak'=>$request->kontak,
-            'deskripsi'=>$request->deskripsi,
-            'alamat'=>$request->alamat,
+            'jumlah'=>$request->jumlah,
+            'biaya'=>$request->biaya,
+            'biaya_transport'=>$request->biaya_transport,
+            'id_mitra_kerja'=>$request->id_mitra_kerja,
+            'status'=>'Menunggu persetujuan',
+            'tanggal'=>$request->tanggal,
         ]);
+        foreach ($request->id_gudang as $id_gudang) {
+            $pengadaan->kegudang()->create([
+                'jumlah'=>$request['isi_gudang_'.$id_gudang],
+                'id_gudang'=>$id_gudang,
+            ]);
+        }
         return redirect()->route('pengadaan.index')->with('success_msg', 'Pengadaan berhasil dibuat');
     }
 
@@ -78,7 +102,17 @@ class PengadaanController extends Controller
      */
     public function show(Pengadaan $pengadaan)
     {
-        //
+        // DB::enableQueryLog();
+        $pengadaan->load('kegudang.gudang');
+        // return DB::getQueryLog();
+        // return $pengadaan;
+        return view('pengadaan.detail', [
+            'title'         => 'Detail Pengadaan',
+            'modul_link'    => route('pengadaan.index'),
+            'modul'         => 'Pengadaan',
+            'active'        => 'pengadaan.index',
+            'd'=>$pengadaan, 
+        ]);
     }
 
     /**
@@ -89,13 +123,16 @@ class PengadaanController extends Controller
      */
     public function edit(Pengadaan $pengadaan)
     {
+        $pengadaan->load('kegudang.gudang');
         return view('pengadaan.ubah', [
             'd'             => $pengadaan,
             'title'         => 'Ubah Pengadaan',
             'modul_link'    => route('pengadaan.index'),
             'modul'         => 'Pengadaan',
             'action'        => route('pengadaan.update', $pengadaan->id),
-            'active'        => 'pengadaan.edit',
+            'active'        => 'pengadaan.index',
+            'listMitraKerja'=>MitraKerja::listMode(),
+            'gudang'=>Gudang::all(),
         ]);
     }
 
@@ -108,19 +145,36 @@ class PengadaanController extends Controller
      */
     public function update(Request $request, Pengadaan $pengadaan)
     {
-        $request->validate([
-            'nama'=>'required|min:3', 
-            'bidang'=>'required|min:3', 
-            'kontak'=>'required|min:3', 
+        $rules = [];
+        $validator = Validator::make($request->all(), [
+            'tanggal'=>'required|date_format:Y-m-d', 
+            'jumlah'=>'required|numeric', 
+            'biaya'=>'required|numeric', 
+            'biaya_transport'=>'required|numeric',
         ]);
+        if(!isset($request->id_gudang)){
+            return redirect()->back()->withErrors($validator)->withInput()->with('error_msg', 'Setidaknya pilih salah satu gudang');
+        }
+        foreach ($request->id_gudang as $id_gudang) {
+            $rules['isi_gudang_'.$id_gudang] = 'required|numeric';
+        }
+        $request->validate($rules);
         $data = [
-            'nama'=>$request->nama,
-            'bidang'=>$request->bidang,
-            'kontak'=>$request->kontak,
-            'deskripsi'=>$request->deskripsi,
-            'alamat'=>$request->alamat,
+            'jumlah'=>$request->jumlah,
+            'biaya'=>$request->biaya,
+            'biaya_transport'=>$request->biaya_transport,
+            'id_mitra_kerja'=>$request->id_mitra_kerja,
+            'status'=>'Menunggu persetujuan',
+            'tanggal'=>$request->tanggal,
         ];
         $pengadaan->update($data);
+        $pengadaan->kegudang()->delete();
+        foreach ($request->id_gudang as $id_gudang) {
+            $pengadaan->kegudang()->create([
+                'jumlah'=>$request['isi_gudang_'.$id_gudang],
+                'id_gudang'=>$id_gudang,
+            ]);
+        }
         return redirect()->route('pengadaan.index')->with('success_msg', 'Pengadaan berhasil diperbarui');
     }
 
@@ -134,5 +188,13 @@ class PengadaanController extends Controller
     {
         $pengadaan->delete();
         return redirect()->back()->with('success_msg', 'Pengadaan berhasil dihapus');
+    }
+
+    public function verifikasi(Pengadaan $pengadaan)
+    {
+        $pengadaan->update([
+            'status'=>'Dalam pengerjaan',
+        ]);
+        return redirect()->back()->with('success_msg', 'Pengadaan berhasil diverifikasi');
     }
 }
